Add MOB_SCALE and move MobType to Model
