+++ server/src/main/java/roguelike/model/util/MobType.java
@@ -1,4 +1,4 @@
-package roguelike.util;
+package roguelike.model.util;
 
 public enum MobType {
     PLAYER,
